+++ src/main/java/com/algaworks/algafood/infrastructure/EstadoRepositoryImpl.java
@@ -1,15 +1,15 @@
 package com.algaworks.algafood.infrastructure;
 
 import java.util.List;
-
-import org.springframework.stereotype.Component;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import com.algaworks.algafood.domain.model.Estado;
 import com.algaworks.algafood.domain.repository.EstadoRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 
-@Component
+@Repository
 public class EstadoRepositoryImpl implements EstadoRepository {
     
     @PersistenceContext
@@ -30,8 +30,13 @@ public class EstadoRepositoryImpl implements EstadoRepository {
     
     @Transactional
     @Override
-    public void remover(Estado estado){
-        estado = buscar(estado.getId());
+    public void remover(Long id){
+        Estado estado = buscar(id);
+
+        if (estado == null) {
+            throw new EmptyResultDataAccessException(1);
+        }
+
         entityManager.remove(estado);
     }
